pull day trips functions into dayTrips component. pull fetch image into helper function
