+++ components/DayTrips.tsx
@@ -1,11 +1,143 @@
-import { ReactElement } from "react";
+import { useEffect, useState } from "react";
 import Page from "./Page";
 import styles from "./day.module.css";
-import { DayTrip, Photo } from "../utils/types";
+import { DayTrip, Meta, Photo, Food } from "../utils/types";
 import { DAY_TRIP_IDS } from "../utils/constants";
+import { getStreamResponse } from "../utils/getStreamResponse";
+import isJsonString from "../utils/isJsonString";
+import fetchImage from "../utils/fetchImage";
 
-export default function DayTrips({ dayTrips, locationName, retry }): ReactElement {
-  if (!dayTrips) return;
+export default function DayTrips({
+  dayTrips,
+  city,
+  meta,
+  getInterestsString,
+  setDayTrips,
+  setMeta }) {
+
+
+  useEffect(() => {
+    fetchDayTripFoods();
+    fetchDayTripDescriptions();
+  }, [meta]);
+
+  async function retry(index: number): Promise<string> {
+
+    const response = await fetch("/api/generateRetryDayTrip", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ city: city, interests: getInterestsString(), currentTrips: meta.dayTrips }),
+    });
+    if (!response.ok) {
+      throw new Error(`Request failed with status ${response.statusText}`);
+    }
+
+    const data = response.body;
+    if (!data) {
+      return;
+    }
+    getStreamResponse(data).then((streamResponse: string) => {
+      setMeta((prevState: Meta): Meta => {
+        const nextState = { ...prevState };
+        nextState.dayTrips[index] = streamResponse;
+        return nextState;
+      });
+
+      setDayTrips((prevState: DayTrip[]): DayTrip[] => {
+        const nextState = [...prevState];
+        nextState[index] = {
+          name: streamResponse,
+          short_desc: "",
+          long_desc: "",
+          food: { name: "", desc: "" },
+          image: {} as Photo
+        };
+
+        return nextState;
+      });
+    });
+  }
+
+  function fetchDayTripDescriptions(): void {
+    for (let i: number = 0; i < dayTrips.length; i++) {
+      fetchDayTripDescription(dayTrips[i], i)
+    }
+  }
+  async function fetchDayTripDescription(location: DayTrip, index: number): Promise<string> {
+    const response = await fetch("/api/generateActivityDescription", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ location: location.name, city: city, interests: getInterestsString() })
+    });
+
+    const responseData = await response.json();
+    if (response.status !== 200) {
+      console.log(response);
+      throw responseData.error || new Error(`Request failed with status ${response.status}`);
+    }
+    const jsonStr: string = JSON.parse(responseData).result;
+    if (!isJsonString(jsonStr)) {
+      return;
+    }
+    const json: DayTrip = JSON.parse(jsonStr);
+    setDayTrips((prevState: DayTrip[]): DayTrip[] => {
+      let updatedDayTrips: DayTrip[] = [...prevState];
+      updatedDayTrips[index].long_desc = json.long_desc;
+      updatedDayTrips[index].short_desc = json.short_desc;
+      return updatedDayTrips;
+    });
+
+    // setLoading((prev: LoadingState): LoadingState => ({
+    //   days: prev.dayTrips,
+    //   dayTrips: false,
+    // }));
+  }
+
+  function fetchDayTripFoods(): void {
+    for (let i: number = 0; i < dayTrips.length; i++) {
+      fetchDayTripFood(dayTrips[i], i);
+    }
+  }
+
+  async function fetchDayTripFood(dayTrip: DayTrip, index: number): Promise<string> {
+    const response = await fetch("/api/generateFood", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json"
+      },
+      body: JSON.stringify({ location: dayTrip.name, city: city }),
+    });
+
+    if (!response.ok) {
+      throw new Error(`Request failed with status ${response.statusText}`);
+    }
+
+    const responseData = await response.json();
+    if (response.status !== 200) {
+      throw responseData.error || new Error(`Request failed with status ${response.status}`);
+    }
+
+    const jsonStr: string = JSON.parse(responseData).result;
+
+    if (!isJsonString(jsonStr)) {
+      return;
+    }
+    const json: Food = JSON.parse(jsonStr);
+
+    fetchImage(dayTrip.name, index).then((image: Photo) => {
+      setDayTrips((prevState: DayTrip[]): DayTrip[] => {
+        const nextState = [...prevState];
+        nextState[index].food = json.lunch;
+        nextState[index].image = image;
+        return nextState;
+      });
+    })
+  }
+
   return dayTrips.map((trip: DayTrip, index: number) => {
     const image: Photo = trip.image;
     let urls = !image ? { regular: '' } : image.urls;
@@ -13,7 +145,7 @@ export default function DayTrips({ dayTrips, locationName, retry }): ReactElemen
     let username: string = !user ? '' : user.username
     return (
       <Page
-        header={locationName}
+        header={city}
         subheader="Day Trip"
         key={index}
         id={DAY_TRIP_IDS[index]}
